fix(tests): use clean_ci_env fixture to clear CI env vars in tests
Create reusable fixture and helper functions for clearing CI environment
variables that trigger non-interactive mode.

Changes:
- Added clean_ci_env fixture in conftest.py
- Added clear_all_ci_env_vars() and restore_env_vars() helpers to conftest
- Included DAF_MOCK_MODE in CI_ENV_VARS list
- Updated test_daf_new_sets_status_paused_after_claude_exits to use fixture
- Updated test_non_interactive_params.py to import helpers from conftest

This ensures tests work correctly in CI/CD environments where CI vars
like GITHUB_ACTIONS, CI, DAF_MOCK_MODE are set.
